requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
click==8.1.7
fastapi==0.112.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.3.2
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
typing_extensions==4.12.2
uvicorn==0.30.6
